what4-domains: Split out abstract domain package
Fixes #370.
