+++ cabal.project
@@ -8,6 +8,7 @@ packages:
   what4/
   what4-abc/
   what4-blt/
+  what4-domains/
   what4-transition-system/
   w4smt2/
