+++ what4-domains/doc/arithdomain.cry
