+++ what4/src/What4/Utils/AbstractDomains.hs
@@ -108,8 +108,8 @@ import           Data.Parameterized.TraversableFC
 import           Data.Ratio (denominator)
 
 import           What4.BaseTypes
-import           What4.Utils.BVDomain (BVDomain)
-import qualified What4.Utils.BVDomain as BVD
+import           What4.Domains.BV (BVDomain)
+import qualified What4.Domains.BV as BVD
 import           What4.Utils.Complex
 import           What4.Utils.StringLiteral
